requirements.txt
asgiref==3.3.2
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-bulma==0.8.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.7.2
python-dateutil==1.5
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
unicorn==1.0.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
